first models created

+++ src/ECommerceService/ECommerceService.Domain/Entities/Base/IEntity.cs
@@ -12,7 +12,7 @@
         public int RowStatus { get; set; }
 
     }
-    public interface IEntity : IEntity<int>
+    public interface IEntity : IEntity<long>
     {
 
     }
